+++ R/run.sensitivity_function.R
@@ -41,6 +41,10 @@
 #'   argument of the \code{\link[R2jags:jags]{jags}} function of the R-package
 #'   \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
 #'   The default argument is 1.
+#' @param inits A list with the initial values for the parameters; an argument
+#'   of the \code{\link[R2jags:jags]{jags}} function of the R-package
+#'   \href{https://CRAN.R-project.org/package=R2jags}{R2jags}.
+#'   The default argument is \code{NULL}, and JAGS generates the initial values.
 #'
 #' @return A list of R2jags outputs on the summaries of the posterior
 #'   distribution, and the Gelman-Rubin convergence diagnostic
@@ -116,6 +120,12 @@
 #'   \code{n_iter}, \code{n_burnin}, and \code{n_thin} as specified by the user
 #'   to be inherited by other relevant functions of the package.
 #'
+#'   The model is updated until convergence using the
+#'   \code{\link[R2jags:autojags]{autojags}} function of the R-package
+#'   \href{https://CRAN.R-project.org/package=R2jags}{R2jags} with 2 updates and
+#'   number of iterations and thinning equal to \code{n_iter} and \code{n_thin},
+#'   respectively.
+#'
 #'   \code{run_sensitivity} can be used only when missing participant
 #'   outcome data have been extracted for at least one trial. Otherwise, the
 #'   execution of the function will be stopped and an error message will be
@@ -123,7 +133,8 @@
 #'
 #' @author {Loukia M. Spineli}
 #'
-#' @seealso \code{\link[R2jags:jags]{jags}}, \code{\link{run_model}}
+#' @seealso \code{\link[R2jags:autojags]{autojags}},
+#'   \code{\link[R2jags:jags]{jags}}, \code{\link{run_model}}
 #'
 #' @references
 #' Gelman, A, Rubin, DB. Inference from iterative simulation using multiple
@@ -179,7 +190,8 @@ run_sensitivity <- function(full,
                             n_chains,
                             n_iter,
                             n_burnin,
-                            n_thin) {
+                            n_thin,
+                            inits = NULL) {
 
 
   if (!inherits(full, "run_model") || is.null(full)) {
@@ -289,6 +301,12 @@ run_sensitivity <- function(full,
   } else {
     n_thin
   }
+  inits <- if (is.null(inits)) {
+    message("JAGS generates initial values for the parameters.")
+    NULL
+  } else {
+    inits
+  }
 
   # A 2x2 matrix of 25 reference-specific scenarios (PMID: 30223064)
   mean_misspar <- as.matrix(cbind(rep(mean_scenarios,
@@ -296,7 +314,7 @@ run_sensitivity <- function(full,
                                   rep(mean_scenarios, length(mean_scenarios))))
 
   # Prepare parameters for JAGS
-  jagsfit <- data_jag <- list()
+  jagsfit0 <- jagsfit <- data_jag <- list()
 
   ## Parameters to save
   param_jags <- if (model == "RE") {
@@ -342,17 +360,25 @@ run_sensitivity <- function(full,
     }
 
     message(paste(i, "out of", length(mean_misspar[, 1]), "total scenarios"))
-    jagsfit[[i]] <- jags(data = data_jag[[i]],
-                         parameters.to.save = param_jags,
-                         model.file =
-                           textConnection(prepare_model(measure,
-                                                        model,
-                                                        covar_assumption = "NO",
-                                                        assumption)),
-                         n.chains = n_chains,
-                         n.iter = n_iter,
-                         n.burnin = n_burnin,
-                         n.thin = n_thin)
+    jagsfit0[[i]] <- jags(data = data_jag[[i]],
+                          inits = inits,
+                          parameters.to.save = param_jags,
+                          model.file =
+                            textConnection(prepare_model(measure,
+                                                         model,
+                                                         covar_assumption ="NO",
+                                                         assumption)),
+                          n.chains = n_chains,
+                          n.iter = n_iter,
+                          n.burnin = n_burnin,
+                          n.thin = n_thin)
+
+    # Update until convergence is necessary
+    message(paste("Updating model for scenario", i, "until convergence"))
+    jagsfit[[i]] <- autojags(jagsfit0[[i]],
+                             n.iter = n_iter,
+                             n.thin = n_thin,
+                             n.update = 2)
   }
 
   # Obtain the posterior distribution of the necessary model paramters