+++ NEWS.md
@@ -1,5 +1,26 @@
-# rnmamod, version 0.4.0 (2023-0X-XX)
+# rnmamod, version 0.4.0 (2024-0X-XX)
 
+ * Function __comp_clustering__: 
+   - Performs quantitative evaluation of the transitivity assumption using
+     inter-study dissimilarities for various study-level aggregate participant 
+     and methodological characteristics that may act as effect modifiers. 
+ * Function __dendro_heatmap__:  
+   - Returns dendrogram with integrated heatmap of the clustered comparisons and
+     studies based on hierarchical agglomerative clustering (performed using the
+     function __comp_clustering__). The R packages _heatmaply_ and _dendextend_ 
+     have been used.
+ * Function __distr_characteristics__:  
+   - It returns violin plots with box plots and dots for quantitative 
+   characteristics, and stacked barplots for qualitative characteristics across
+   the observed treatment comparisons. The function can  be also used to 
+   illustrate the distribution of the characteristics across the clusters 
+   defined from __comp_clustering__.
+ * Function __miss_characteristics__:  
+   - It returns various plots to visualise the missing cases in the 
+   characteristics across studies and treatment comparisons.
+ * Function __gower_distance__: 
+   - It returns the N-by-N matrix on Gower's dissimilarity coefficient for all
+   pairs of N studies in a network.
  * Function __mcmc_diagnostics__: 
    - returns a barplot on the ratio of MCMC error to the posterior standard
    deviation and a barplot on the Gelman-Rubin R diagnostic. Green bars